Add files via upload
Geofencing with Task Management is an integrated intranet system, developed for use by the faculty personnel within college, universities or offices, adapted to Smartphone's, tablets and handheld devices with the campus or university under the same LAN. Most of the colleges in India do traditional manual paperwork to maintain the records of teacher’s attendance & schedule which become burdensome. Though it works efficiently most of the time but compromises time and space. The proposed system which is based on an Android platform facilitates principal, HOD and the personnel's faculty staff throughout with the facility, regardless of the existence of network connection in the area using a typical smartphone. The proposed application and its backend system maintains access to teachers location based on Real-Time Location System (RTLS), called geofencing as well as maintaining & manipulating their timetable, checking out today’s schedule of a teacher. Additional features include editing the timetable for minor changes or import excel sheet for major changes. Also, we integrate a Quick Response (QR code) for checking the availability of the classroom using a smartphone which will help to maintain a google spreadsheet of classrooms. Overall this project of ours is being developed to help the teachers as well as Head of Department to maintain their daily schedule in the best way possible.
For better understanding, we have uploaded our black book references and presentation for visual understanding.